Point2f->Point2d and simplification.  This avoids the rounding errors that used to break the regression tests

+++ include/cosmosim/PixMap.h
@@ -28,7 +28,7 @@ public:
     using LensModel::LensModel ;
 
 protected:
-    virtual cv::Point2f getDistortedPos(double r, double theta) const;
+    virtual cv::Point2d getDistortedPos(double r, double theta) const;
     virtual void updateApparentAbs() ;
 
 };
