+++ src/simlib/simaux.h
@@ -8,8 +8,8 @@ double factorial_(unsigned int) ;
 void diffX(cv::InputArray, cv::OutputArray ) ;
 void diffY(cv::InputArray, cv::OutputArray ) ;
 
-cv::Point2f pointCoordinate( cv::Point2f pt, cv::Mat im ) ;
-cv::Point2f imageCoordinate( cv::Point2f pt, cv::Mat im ) ;
+cv::Point2d pointCoordinate( cv::Point2d pt, cv::Mat im ) ;
+cv::Point2d imageCoordinate( cv::Point2d pt, cv::Mat im ) ;
 void gradient(cv::InputArray, cv::OutputArray, cv::OutputArray ) ;
 // #define diffX(src,out) Sobel(src,out,CV_64FC1, 1, 0, 3, 1.0/8)
 // #define diffY(src,out) Sobel(src,out,CV_64FC1, 0, 1, 3, -1.0/8)
